ENGR00154050 UMS: The sd card is full after erase, have to format on PC. 100%
It is the limitation of using FAT32 in windows. A FAT32 volume in windows
must contain a minimum of 65,527 clusters. In this case the clusters was
less than this number.
Fixed the bug by change the clusters to a small number. When clusters are
too small, use FAT16 instead or FAT12 when the clusters are very small.

Signed-off-by: YangYonggang <b31664@freescale.com>