init: rework kernel command-line fstab override mechanism
As the cmdline parsing was done in do_mount_all, it wasn't called in
recovery console as partitions are not mounted by default.
Also, as the env variables weren't added to the ENV array, the forked
processes (such as /sbin/recovery) were missing the blockdev variable
mandatory for mounting partitions.
Finally, this export to env is now limited to androidboot* variables due
to the size of ENV array.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>